Added functionality to create wim images.
Modified a PostProcessor function return directive and moved creation of reutrn object to body of function.
Changed some declaration of variables used in code to be make it more readable.
Added artifact.go that hold struct that implements packer.Artifact interface and is used in reutrn of PostProcessor.
Added unmounting of VHD/VHDX image in case of cancelation while WIM is created by DISM.